07_revenue_data_extraction/requirements.txt
pymupdf>=1.24.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0
azure-identity>=1.16.0
azure-storage-blob>=12.20.0
python-dotenv>=1.0.0
pandas>=2.2.0
pydantic>=2.7.0
tiktoken>=0.7.0
